SDP/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
Pillow>=10.0.0
python-decouple>=3.8
python-dotenv>=1.0.0

# Additional packages that might be useful
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0

# For production
gunicorn>=21.0.0
whitenoise>=6.0.0

# For development
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0 